Bunch of portability related changes to xfsprogs.  Includes some code to allow db/mkfs/repair to be used on Mac's (maybe useful for dual boot), & some code to build on IRIX too (useful for purify-ing xfs_repair only). bump version for portability changes.

+++ db/check.c
@@ -2674,8 +2674,8 @@ process_inode(
 		dbprintf("inode %lld mode %#o fmt %s "
 			 "afmt %s "
 			 "nex %d anex %d nblk %lld sz %lld%s%s\n",
-			id->ino, dic->di_mode, fmtnames[dic->di_format],
-			fmtnames[dic->di_aformat],
+			id->ino, dic->di_mode, fmtnames[(int)dic->di_format],
+			fmtnames[(int)dic->di_aformat],
 			dic->di_nextents,
 			dic->di_anextents,
 			dic->di_nblocks, dic->di_size,
